+++ src/utils/get_random.js
@@ -1,0 +1,17 @@
+export const getRandomInteger = (a = 0, b = 1) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+
+  return Math.floor(lower + Math.random() * (upper - lower + 1));
+};
+
+//генерирует случайный элемент массива
+export const generateRandom = (arr) => {
+  const randomIndex = getRandomInteger(0, arr.length - 1);
+  return arr[randomIndex];
+};
+
+//генерирует случайное число до максимального
+export const getRandomNumber = (min, max) => {
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+};
